Change tabs into spaces and rename the chatbot to Duke
