+++ src/main/java/EmptyListException.java
@@ -1,5 +1,8 @@
+/**
+ * Exception class that indicates the list is empty
+ */
 public class EmptyListException extends Exception {
-	public EmptyListException(String errorMessage) {
-		super(errorMessage);
-	}
+    public EmptyListException(String errorMessage) {
+        super(errorMessage);
+    }
 }
