+++ src/main/java/InvalidCommandException.java
@@ -1,5 +1,8 @@
+/**
+ * Exception class that indicates the input command doesn't exist
+ */
 public class InvalidCommandException extends Exception {
-	public InvalidCommandException(String errorMessage) {
-		super(errorMessage);
-	}
+    public InvalidCommandException(String errorMessage) {
+        super(errorMessage);
+    }
 }
