+++ src/main/java/InvalidCommandSyntaxException.java
@@ -1,5 +1,8 @@
+/**
+ * Exception class that indicates the command is typed wrongly
+ */
 public class InvalidCommandSyntaxException extends Exception {
-	public InvalidCommandSyntaxException(String errorMessage) {
-		super(errorMessage);
-	}
+    public InvalidCommandSyntaxException(String errorMessage) {
+        super(errorMessage);
+    }
 }
