+++ src/main/java/InvalidTaskNumberException.java
@@ -1,5 +1,8 @@
+/**
+ * Exception class that indicates the task number is out of bounds
+ */
 public class InvalidTaskNumberException extends Exception {
-	public InvalidTaskNumberException(String errorMessage) {
-		super(errorMessage);
-	}
+    public InvalidTaskNumberException(String errorMessage) {
+        super(errorMessage);
+    }
 }
