+++ src/main/java/InvalidTaskTimeException.java
@@ -1,5 +1,8 @@
+/**
+ * Exception class that indicates there is missing information on time inputs
+ */
 public class InvalidTaskTimeException extends Exception {
-	public InvalidTaskTimeException(String errorMessage) {
-		super(errorMessage);
-	}
+    public InvalidTaskTimeException(String errorMessage) {
+        super(errorMessage);
+    }
 }
