+++ src/main/java/MissingTaskDescriptionException.java
@@ -1,5 +1,8 @@
+/**
+ * Exception class that indicates the task is missing its description
+ */
 public class MissingTaskDescriptionException extends Exception {
-	public MissingTaskDescriptionException(String errorMessage) {
-		super(errorMessage);
-	}
+    public MissingTaskDescriptionException(String errorMessage) {
+        super(errorMessage);
+    }
 }
