+++ src/main/java/MissingTaskNameException.java
@@ -1,5 +1,8 @@
+/**
+ * Exception class that indicates the task has no name
+ */
 public class MissingTaskNameException extends Exception {
-	public MissingTaskNameException(String errorMessage) {
-		super(errorMessage);
-	}
+    public MissingTaskNameException(String errorMessage) {
+        super(errorMessage);
+    }
 }
