+++ src/main/java/MissingTaskNumberException.java
@@ -1,5 +1,8 @@
+/**
+ * Exception class that indicates no number was inputted
+ */
 public class MissingTaskNumberException extends Exception {
-	public MissingTaskNumberException(String errorMessage) {
-		super(errorMessage);
-	}
+    public MissingTaskNumberException(String errorMessage) {
+        super(errorMessage);
+    }
 }
