+++ src/main/java/ToDo.java
@@ -2,19 +2,29 @@ import java.util.ArrayList;
 
 /**
  * Class to declare a Todo task
- * 
- * @author Jason Ray
  */
 public class ToDo extends Task {
-	public ToDo(String name, Type type) {
-		super(name, type, "");
-	}
+    /**
+     * Creates a new instance of a ToDo task
+     * 
+     * @param name Name of task
+     */
+    public ToDo(String name) {
+        super(name, Type.TODO, "");
+    }
 
-	public static void printTaskAdded(String taskName, ArrayList<Task> list) {
-		System.out.println("You have added a task:");
-		System.out.println("\t[T][ ] " + taskName);
-		list.add(new ToDo(taskName, Type.TODO));
-		System.out.println("There are now " + list.size() + " tasks in the list");
-	}
+    /**
+     * Prints out the confirmation text output when a todo task is added
+     * Also adds the task to the list
+     * 
+     * @param taskName Name of task
+     * @param list List where task is to be added
+     */
+    public static void printTaskAdded(String taskName, ArrayList<Task> list) {
+        System.out.println("You have added a task:");
+        System.out.println("\t[T][ ] " + taskName);
+        list.add(new ToDo(taskName));
+        System.out.println("There are now " + list.size() + " tasks in the list");
+    }
 
 }
