+++ src/main/java/Type.java
@@ -1,5 +1,8 @@
+/**
+ * Enum class to declare types of tasks
+ */
 public enum Type {
-	TODO,
-	DEADLINE,
-	EVENT;
+    TODO,
+    DEADLINE,
+    EVENT;
 }
